+++ tests/integration/cases/empty_sequence/VisualBasic_default_sequence_element_type_string.vb
@@ -1,0 +1,7 @@
+Imports System.Collections.Generic
+Module Check
+    Dim my_data = New Object() {
+        New String() {},
+        New Dictionary(Of String, Object) From {}
+    }
+End Module